requirements.txt
streamlit==1.32.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38
openai==1.51.0
pypdf2==3.0.1
python-dotenv==1.0.1